# external/interface/httpintf/swagger/api.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to provide API over HTTP to toggler service,\nin which you can interact with the service in a programmatic way.",
    "title": "provides API on HTTP layer to the toggler service.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "version": "0.2.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/deployment-environments/{envID}/release-flags/{flagID}/release-rollouts": {
      "get": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "rollout",
          "flag"
        ],
        "summary": "List all the release flag that can be used to manage a feature rollout.",
        "operationId": "listReleaseRollouts",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "EnvironmentID",
            "description": "EnvironmentID is the deployment environment ID",
            "name": "envID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "FlagID",
            "description": "FlagID is the release flag id",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listReleaseRolloutResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "rollout",
          "flag"
        ],
        "summary": "This operation allows you to create a new release rollout.",
        "operationId": "createReleaseRollout",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "EnvironmentID",
            "description": "EnvironmentID is the deployment environment ID",
            "name": "envID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "FlagID",
            "description": "FlagID is the release flag id",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rollout": {
                  "type": "object",
                  "required": [
                    "plan"
                  ],
                  "properties": {
                    "plan": {
                      "description": "Plan holds the composited rule set about the pilot participation decision logic.",
                      "type": "object",
                      "x-go-name": "Plan",
                      "example": {
                        "percentage": {
                          "percentage": 42,
                          "seed": 10240
                        }
                      }
                    }
                  },
                  "x-go-name": "Rollout"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createReleaseRolloutResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/deployment-environments/{envID}/release-flags/{flagID}/release-rollouts/{rolloutID}": {
      "put": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "rollout",
          "flag",
          "pilot"
        ],
        "summary": "Update a release flag.",
        "operationId": "updateReleaseRollout",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "EnvironmentID",
            "description": "EnvironmentID is the deployment environment ID",
            "name": "envID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "FlagID",
            "description": "FlagID is the release flag id",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "RolloutID",
            "description": "FlagID is the release flag id or the alias name.",
            "name": "rolloutID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rollout": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "object",
                      "x-go-name": "Plan"
                    }
                  },
                  "x-go-name": "Rollout"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateReleaseRolloutResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "description": "Error contains the details of the error",
      "type": "object",
      "properties": {
        "code": {
          "description": "The constant code of the error that can be used for localisation",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 401
        },
        "message": {
          "description": "The message that describe the error to the developer who do the integration.\nNot meant to be propagated to the end-user.\nThe Message may change in the future, it it helps readability,\nplease do not rely on the content in any way other than just reading it.",
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/external/interface/httpintf/httpapi"
    },
...
    "Rollout": {
      "type": "object",
      "properties": {
        "env_id": {
          "type": "string",
          "x-go-name": "EnvironmentID"
        },
        "flag_id": {
          "type": "string",
          "x-go-name": "FlagID"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "plan": {
          "type": "object",
          "x-go-name": "Plan"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/external/interface/httpintf/httpapi"
    }
  },
  "responses": {
...
    "createReleaseRolloutResponse": {
      "description": "CreateReleaseRolloutResponse",
      "schema": {
        "type": "object",
        "properties": {
          "rollout": {
            "$ref": "#/definitions/Rollout"
          }
        }
      }
    },
    "errorResponse": {
      "description": "ErrorResponse will contains a response about request that had some kind of problem.\nThe details will be included in the body.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
...
    "listReleaseRolloutResponse": {
      "description": "ListReleaseRolloutResponse",
      "schema": {
        "type": "object",
        "properties": {
          "rollouts": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Rollout"
            },
            "x-go-name": "Rollouts"
          }
        }
      }
    },
...
    "updateReleaseRolloutResponse": {
      "description": "UpdateReleaseRolloutResponse",
      "schema": {
        "type": "object",
        "properties": {
          "rollout": {
            "type": "object",
            "properties": {
              "plan": {
                "type": "object",
                "x-go-name": "Plan"
              }
            },
            "x-go-name": "Rollout"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "AppKey": {
      "type": "apiKey",
      "name": "X-APP-KEY",
      "in": "header"
    },
    "AppToken": {
      "type": "apiKey",
      "name": "X-APP-TOKEN",
      "in": "header"
    }
  },
  "security": [
    {
      "AppKey": [
        "[]"
      ]
    }
  ]
}